Update MenuFunctions.sh
-- LINUX DIRECTORY BACKUP SHELL SCRIPT --

INTRODUCTION

This project was undertook to develop an understanding in using Linux CLI, using Ubuntu. To understand typical and basic operations of a System Administrator and systems networking. I implemented a small back system that enable users (employees) to login, edit files, make transfers (intranet to live) manually, automatically backup and log files and perform scheduled backups. This entire system was develop solely on the Command Line Terminal, using Shell Scripting (bash) and Virtual Box for virtualization.

The system provides 2 users (employee) and a user group (employees). It also holds 2 directories, Live and Intranet, and files that users can edit and modify (CSS and Javascript). This system provides a menu for users to navigate, edit, transfer and log updates they have made. It also records and logs the actions users have made and logs this to an audit log file to trace actions. The system also backups the entire Intranet directory to the Live directory each day at a specified scheduled time, using Crontab. In information technology, a backup, or data backup is a copy of computer data taken and stored elsewhere so that it may be used to restore the original after a data loss event

This system incorporates difference user permissions and locks down any edits from being performed during the scheduled transfer.


SHELL SCRIPTING (Bash)

Bash (AKA Bourne Again Shell) is a type of interpreter that processes shell commands. A shell interpreter takes commands in plain text format and calls Operating System services to do something. A shell scripting is writing a program for the shell to execute and a shell script is a file or program that shell will execute. A shell script is a fully-fledged programming language in itself. It can define variables, functions and we can do conditional execution of shell commands as well.

In this project I created a menu to allow users to navigate and perform actions.  This is implemented by first developing the specific statements and labelling with a number. Then using an if/else clause to determine which function should be called. The menu is continuously displayed using a while loop and clear function. The separate functions have individual responsibility in performing their tasks. This is to keep the code clean and robust.


VIRTUALIZATION

I implemented this project using Virtualization to understand its fundamental, efficiency and use in real world tasks. 

Virtualization - uses software to create an abstraction layer over computer hardware that allows the hardware elements of a single computer—processors, memory, storage and more—to be divided into multiple virtual computers, commonly called virtual machines (VMs). Each VM runs its own operating system (OS) and behaves like an independent computer, even though it is running on just a portion of the actual underlying computer hardware. It benefits resource efficiency, easier management, minimal downtime and faster provisioning.