Update MenuFunctions.sh
-- LINUX DIRECTORY BACKUP SHELL SCRIPT --

INTRODUCTION

This project was to develop an understanding in using Linux CLI, Ubuntu. To understand typical and basic operations of a system administrator and systems networking. Familiarise myself with a Linux distro and virtualisation.
I implemented a small Linux admin system that enable users (employees) to login, edit files, make file transfers (intranet to live) manually, automatically backup as well log files. The system also included a scheduled backup feature. This entire system was developed solely on the Command Line Interpreter, using Bash Shell and Virtual Box for virtualization.

The system provides users and a user groups. It also holds 2 directories, Live and Intranet, and files that users can edit and modify (CSS and Javascript). This system provides a menu for users to navigate, edit, transfer and log updates they have made. It also records and logs the actions users have made and logs this to an audit log file to track actions (date, time and ID). The system also backups the entire Intranet directory to the Live directory each day at a specified scheduled time, using Crontab. In information technology, a backup, or data backup is a copy of computer data taken and stored elsewhere so that it may be used to restore the original after a data loss event.

In this project I created a menu to allow users to navigate and perform actions.  This is implemented by first developing the specific statements and labelling with a number. Then using an if/else clause to determine which function should be called. The menu is continuously displayed using a while loop and clear function. The separate functions have individual responsibility in performing their tasks. This is to keep the code clean and robust.

This system incorporates difference user permissions and locks down any edits from being performed during the scheduled transfer. The permissions are similarly designed to keep users from modifying whole system features. These permissions are kept for the administrator. 


SHELL SCRIPTING (Bash)

Bash (AKA Bourne Again Shell) is an interpreter that processes shell commands. A shell interpreter takes commands in plain text format and calls the Operating System services to do something. A shell scripting is writing a program for the shell to execute and a shell script is a file or program that shell will execute. A shell script is a fully-fledged programming language in itself. It can define variables, functions and we can do conditional execution of shell commands as well. Bash is the most common and most intuitive scripting language for Linux distros.


VIRTUALIZATION

This application was designed on  virtual machine using Virtual Box. 

Virtualization - uses software to create an abstraction layer over computer hardware that allows the hardware elements of a single computer—processors, memory, storage and more—to be divided into multiple virtual computers, commonly called virtual machines (VMs). Each VM runs its own operating system (OS) and behaves like an independent computer, even though it is running on just a portion of the actual underlying computer hardware. It benefits resource efficiency, easier management, minimal downtime and faster provisioning.